requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
distro==1.9.0
dnspython==2.4.2
exceptiongroup==1.2.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
load-dotenv==0.1.0
multidict==6.0.4
openai==1.7.0
pydantic==2.5.3
pydantic_core==2.14.6
pymongo==4.6.1
pyTelegramBotAPI==4.14.1
python-dotenv==1.0.0
python-telegram-bot==20.7
requests==2.31.0
sniffio==1.3.0
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==1.26.15
yarl==1.9.4
